+++ projects/gloo/pkg/plugins/upstreamssl/plugin.go
@@ -1,4 +1,4 @@
-package upstream
+package upstreamssl
 
 import (
 	envoyapi "github.com/envoyproxy/go-control-plane/envoy/api/v2"
