+++ projects/gloo/pkg/plugins/upstreamssl/plugin_test.go
@@ -1,4 +1,4 @@
-package upstream_test
+package upstreamssl_test
 
 import (
 	envoyapi "github.com/envoyproxy/go-control-plane/envoy/api/v2"
