+++ projects/gloo/pkg/plugins/upstreamssl/upstream_suite_test.go
@@ -1,4 +1,4 @@
-package upstream_test
+package upstreamssl_test
 
 import (
 	"testing"
